Used StringBuilder to concat strings; Unused imports removed; Const expression used for constant values

+++ app/src/main/java/com/pedro/rtpstreamer/utils/ImageAdapter.java
@@ -32,7 +32,7 @@ public class ImageAdapter extends BaseAdapter {
   }
 
   public long getItemId(int position) {
-    return (long) position;
+    return position;
   }
 
   @Override
